second part undone

+++ lib/components/constants/mediaquares.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
 
 num m_h(context){
-  return MediaQuery.of(context).size.height;
+  return MediaQuery.of(context).size.height.toDouble();
 }
 num m_w(context){
-  return MediaQuery.of(context).size.width;
+  return MediaQuery.of(context).size.width.toDouble();
 }
